Fix PyTorchImageClassificationTrainer's training (#3339)
* Fix PyTorchImageClassificationTrainer's training

It seems that the current process only calculates the loss and gradient, and does not use the optimizer for optimization. Therefore, the model is not actually trained, and its accuracy on the Web UI remains unchanged.

* Add intermediate reports for every epoch

For now, intermediate reports and final reports are consistent, and they are displayed once after all epochs have finished. This may not meet our expectations. We hope that intermediate reports can provide us with the validation results after each epoch.