Fix the bug where the valid_indices array is used as a truth value instead of checking if it has any entries.
When using BOHB, there will be an error as soon as the valid_indices array has more than one entry due to the numpy error 'The truth value of an array with more than one element is ambiguous'. The original BOHB code also uses len(valid_indices) so I changed it accordingly.